com/tm/python/langchain/requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
python-dotenv
pydantic-settings
faiss-cpu
happybase
requests
thriftpy2
ply
six